erlte (erlte v0.1.0)

Types

compiled/0

-type compiled() :: #erlte_compiled{}.

format_specs/0

-type format_specs() :: #erlte_format_specs{}.

node_specs/0

-type node_specs() :: #erlte_node_specs{}.

tags_specs/0

-type tags_specs() :: #erlte_tags{}.

trim_specs/0

-type trim_specs() :: #erlte_trim{}.

Functions

compile(Input)

-spec compile(Input) -> {ok, Compiled} | {error, Error}
           when
               Input :: binary() | list() | {file, file:name_all()},
               Compiled :: erlte:compiled(),
               Error :: term().

Compile the specified input with no variables and default format specifications

compile(Input, Variables)

-spec compile(Input, Variables) -> {ok, Compiled} | {error, Error}
           when
               Input :: binary() | list() | {file, file:name_all()},
               Variables :: list(),
               Compiled :: erlte:compiled(),
               Error :: term().

Compile the specified input with variables and default format specifications

compile(Input, Variables, FormatSpecs)

-spec compile(Input, Variables, FormatSpecs) -> {ok, Compiled} | {error, Error}
           when
               Input :: binary() | list() | {file, file:name_all()},
               Variables :: list(),
               FormatSpecs :: undefined | erlte:format_specs(),
               Compiled :: erlte:compiled(),
               Error :: term().

Compile the specified input according to the format specification

compiled_read_file(Filename)

-spec compiled_read_file(Filename) -> {ok, Compiled} | {error, Reason}
                      when
                          Filename :: file:name_all(),
                          Compiled :: erlte:compiled(),
                          Reason :: file:posix() | badarg | terminated | system_limit.

Read the compiled template from the specified file

compiled_write_file(Filename, Compiled)

-spec compiled_write_file(Filename, Compiled) -> ok | {error, Reason}
                       when
                           Filename :: file:name_all(),
                           Compiled :: erlte:compiled(),
                           Reason :: file:posix() | badarg | terminated | system_limit.

Write the compiled template to the specified file

render(Compiled)

-spec render(Compiled) -> {ok, Result} | {error, Error}
          when Compiled :: erlte:compiled(), Result :: binary(), Error :: term().

Render the compiled template without replacing any variables There is no error return case now but some maybe added in the future so for compatibility you should handle a generic error case too.

render(Compiled, Variables)

-spec render(Compiled, Variables) -> {ok, Result} | {error, Error}
          when
              Compiled :: erlte:compiled(),
              Variables :: list(),
              Result :: binary(),
              Error :: term().

Render compiled template replacing the specified variables There is no error return case now but some maybe added in the future so for compatibility you should handle a generic error case too.

render_to_compiled(Compiled, Variables)

-spec render_to_compiled(Compiled, Variables) -> {ok, NewCompiled} | {error, Error}
                      when
                          Compiled :: erlte:compiled(),
                          Variables :: list(),
                          NewCompiled :: erlte:compiled(),
                          Error :: term().

Render the compiled template replacing the specified variables and return a new compiled template There is no error return case now but some maybe added in the future so for compatibility reasons you should handle a generic error case too.

render_to_file(Filename, Compiled)

-spec render_to_file(Filename, Compiled) -> {ok, Result} | {error, Reason}
                  when
                      Filename :: file:name_all(),
                      Compiled :: erlte:compiled(),
                      Result :: binary(),
                      Reason :: file:posix() | badarg | terminated | system_limit.

Render compiled template and save the result into the specified file

render_to_file(Filename, Compiled, Variables)

-spec render_to_file(Filename, Compiled, Variables) -> {ok, Result} | {error, Reason}
                  when
                      Filename :: file:name_all(),
                      Compiled :: erlte:compiled(),
                      Variables :: undefined | list(),
                      Result :: binary(),
                      Reason :: file:posix() | badarg | terminated | system_limit.

Render compiled template replacing the specified variables and save the result into the specified file