erlte_utils (erlte v0.1.0)

Functions

html_entities_encode(Value)

Simple version of html entities encode

list_remove_start_list(TT, IT)

Remove a list from the begining of another list. If the list is not found false is returned. If the is found the remaining of the second list is returned.

variable_name_to_binary(Name)

Convert variable name to binary